feat(core): fractional partial count — smooth "Partials" sweep
`set_partial_limit` is now `f32` (Voice / PolySynth / C-ABI): the integer
part is the hard cap, the fractional part fades the next partial in
continuously, so the control is smooth across its whole 1.0..2048.0
range — no "steps" even at 1→2→3.

    S_{n+frac}(p) = S_n(p) + frac · r^{n+1} · cos(2π(n+1)p)

— exactly the next member of the finite sum (kernel test
`frac_partial_at_one_equals_the_next_integer_partial` proves it), one
extra `cos`, still Θ(log n). The fraction is suppressed when the Nyquist
clamp is what binds, so it never aliases. Peak normalisation uses the
integer-n peak, so level ripples <1 dB as frac sweeps 0→1 — deliberate,
continuous, documented.

`geometric_partials_pre_frac` in kernel.rs; `Voice::geom_osc` gates on
`frac != 0.0` so the default (2048.0) and Nyquist-bound cases take
`geometric_partials_pre` verbatim — bit-identical, cross-platform hash
unchanged (re-verified: aarch64 + armv7-hf both 76/76, delta 0.0).

C-ABI `harmonic_voice_set_partial_limit` param changed `unsigned int` →
`float` (shipped one commit ago, no external consumers).

Voice: 520 → 528 bytes (+f32, past an 8-byte boundary). Tests 74 → 76.
clippy clean (std + no_std), pluginval strictness 8 + clap-validator
35/35 unchanged. Docs: 01/03/04/05/06/07/08/09/12/README.

+++ harmonic_core/include/harmonic_core.h
@@ -78,11 +78,13 @@ void harmonic_voice_set_hq(HarmonicVoice *voice, unsigned int hq);
  * Unknown values fall back to geometric. */
 void harmonic_voice_set_waveform(HarmonicVoice *voice, unsigned int waveform);
 
-/* Upper bound on the geometric oscillator's partial count -> [1, 2048].
- * Still capped at the Nyquist limit, so lowering it only darkens the tone
- * (never aliases) and costs the same at any setting. 2048 (default) = no
- * limit. Ignored by the sawtooth / triangle waveforms. */
-void harmonic_voice_set_partial_limit(HarmonicVoice *voice, unsigned int limit);
+/* Upper bound on the geometric oscillator's partial count -> [1.0, 2048.0],
+ * fractional: the integer part is the hard cap, the fraction fades the next
+ * partial in for a smooth (non-stepped) sweep. Still capped at the Nyquist
+ * limit, so lowering it only darkens the tone (never aliases) and costs the
+ * same at any setting. 2048.0 (default) = no limit. Ignored by the sawtooth /
+ * triangle waveforms. */
+void harmonic_voice_set_partial_limit(HarmonicVoice *voice, float limit);
 
 /* Reset phase + smoothers + filter state (call on note-on). Honors
  * free_running: in that mode the oscillator phase keeps running. */
